Adding in new assets and modifying map
I added in the new assets I've developed. There is no skin applied. I've also moved some of the Appliance placeholders as I have no clue on what to do with them. The assets are placed in a parent called "1st or 2nd floor Assets". It should be placed with the same layout as the other placeholders and objects.